code clean up, knowledge folder re-organized, and imports fixed
