+++ examples/knowledge_tuning/knowledge_tuning_with_reasoning_model/prompts/generate_summary.yaml
